WaitForMigrationCommandBaseHandler
