+++ src/OctoshiftCLI.Tests/Octoshift/Commands/WaitForMigrationCommandBaseTests.cs
@@ -14,7 +14,7 @@ public class WaitForMigrationCommandBaseTests
     private readonly Mock<ITargetGithubApiFactory> _mockTargetGithubApiFactory = new();
     private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
 
-    private readonly WaitForMigrationCommandBase _command;
+    private readonly WaitForMigrationCommandBaseHandler _command;
 
     private const string REPO_MIGRATION_ID = "RM_MIGRATION_ID";
     private const string ORG_MIGRATION_ID = "OM_MIGRATION_ID";
@@ -23,7 +23,7 @@ public class WaitForMigrationCommandBaseTests
 
     public WaitForMigrationCommandBaseTests()
     {
-        _command = new WaitForMigrationCommandBase(_mockOctoLogger.Object, _mockTargetGithubApiFactory.Object)
+        _command = new WaitForMigrationCommandBaseHandler(_mockOctoLogger.Object, _mockTargetGithubApiFactory.Object)
         {
             WaitIntervalInSeconds = WAIT_INTERVAL
         };
